Improved the RAG system by integrating Context Awareness Gate for retrieval where necessary.
