+++ credit_decision_system/rag/contextual_rag.py
@@ -1,18 +1,60 @@
 from sentence_transformers import SentenceTransformer
-import json
-import faiss
+from context_awareness_gate import ContextAwarenessGate
 from datetime import datetime
+from sklearn.metrics.pairwise import cosine_similarity
+import numpy as np
+import faiss
+import json
 
 class ContextualRAG:
     """
-    Retrieval-Augmented Generation system to determine economic context
+    Retrieval-Augmented Generation system with Context Awareness Gate (CAG)
+    to enhance efficiency and avoid unnecessary retrievals.
     """
     def __init__(self, use_llm=True, client=None, NLP_AVAILABLE=True):
         self.use_llm = use_llm and client is not None
         self.client = client
-        
-        # Define economic context documents (bleak, neutral, positive)
-        self.documents = {
+        self.documents = self._load_documents()
+
+        if NLP_AVAILABLE:
+            try:
+                # Load embedding model
+                self.embedder = SentenceTransformer("all-MiniLM-L6-v2")
+
+                # Prepare document embeddings
+                all_docs, all_contexts = [], []
+                for context, docs in self.documents.items():
+                    all_docs.extend(docs)
+                    all_contexts.extend([context] * len(docs))
+
+                self.doc_embeddings = self.embedder.encode(all_docs, convert_to_numpy=True)
+                self.contexts = all_contexts
+
+                # Build FAISS index for fast retrieval
+                dimension = self.doc_embeddings.shape[1]
+                self.index = faiss.IndexFlatL2(dimension)
+                self.index.add(self.doc_embeddings)
+
+                # Initialize Context Awareness Gate
+                self.cag = ContextAwarenessGate(
+                    embedder=self.embedder,
+                    documents=self.documents,
+                    threshold=0.35,
+                    top_k=3,
+                    use_llm_signal=True,
+                    llm_client=self.client,
+                    debug=self.debug
+                )
+
+                print("✅ ContextualRAG initialized with CAG")
+            except Exception as e:
+                print(f"❌ Error initializing RAG: {e}")
+                self.use_llm = False
+        else:
+            print("⚠️ NLP not available. Falling back to rule-based classification.")
+
+    def _load_documents(self):
+        return {
             "bleak": [
                 "Singapore's tech industry is facing headwinds with over 5,000 layoffs reported in Q2 alone.",
                 "Venture capital funding in Southeast Asia has dropped 40% compared to the previous year.",
@@ -50,132 +92,122 @@ class ContextualRAG:
                 "SMEs in the software sector report excellent cash flow and easy access to credit and capital."
             ]
         }
-        
-        # Set up embedding model if available
-        if NLP_AVAILABLE:
-            try:
-                self.embedder = SentenceTransformer("all-MiniLM-L6-v2")
-                
-                # Flatten all documents with context labels
-                all_docs = []
-                all_contexts = []
-                for context, docs in self.documents.items():
-                    all_docs.extend(docs)
-                    all_contexts.extend([context] * len(docs))
-                
-                # Create embeddings
-                self.doc_embeddings = self.embedder.encode(all_docs, convert_to_numpy=True)
-                self.contexts = all_contexts
-                
-                # Set up FAISS index
-                dimension = self.doc_embeddings.shape[1]
-                self.index = faiss.IndexFlatL2(dimension)
-                self.index.add(self.doc_embeddings)
-                
-                print("✅ RAG system initialized successfully")
-            except Exception as e:
-                print(f"❌ Error setting up RAG: {e}")
-                self.use_llm = False
-        else:
-            print("⚠️ Using rule-based context classification (RAG unavailable)")
-    
+
     def retrieve_context(self, query, top_k=3):
-        """Retrieve most relevant documents for a query"""
-        
+        """Retrieve documents most relevant to query"""
+        expanded_query = f"{query} Related terms: market, finance, investment, tech, economy"
         query_embedding = self.embedder.encode([query], convert_to_numpy=True)
         distances, indices = self.index.search(query_embedding, top_k)
-        
-        retrieved_docs = [self.doc_embeddings[i] for i in indices[0]]
+
         retrieved_contexts = [self.contexts[i] for i in indices[0]]
-        
-        # Count contexts to find most common
-        context_counts = {}
-        for context in retrieved_contexts:
-            context_counts[context] = context_counts.get(context, 0) + 1
-        
-        # Get majority context
-        majority_context = max(context_counts.items(), key=lambda x: x[1])[0]
-        
-        # Get the documents corresponding to the majority context
-        result_docs = [self.documents[majority_context][i % len(self.documents[majority_context])] 
-                      for i, ctx in enumerate(retrieved_contexts) if ctx == majority_context]
-        
+        context_counts = {ctx: retrieved_contexts.count(ctx) for ctx in set(retrieved_contexts)}
+        majority_context = max(context_counts, key=context_counts.get)
+
+        result_docs = [self.documents[majority_context][i % len(self.documents[majority_context])]
+                       for i, ctx in enumerate(retrieved_contexts) if ctx == majority_context]
+
         return result_docs, majority_context
-    
+
     def classify_context_with_llm(self, query):
-        """Use LLM to classify context based on retrieved documents"""
+        """Use LLM to classify economic context based on retrieved documents"""
         if not self.use_llm:
             return self.classify_context_rule_based(query)
-        
+
         try:
             retrieved_docs, majority_context = self.retrieve_context(query)
             context = "\n".join(retrieved_docs)
-            
+
             prompt = f"""
             You are a credit policy assistant. Based on the following macroeconomic context, 
             classify the economic outlook as one of: "bleak", "neutral", or "positive".
-            
+
             Return your answer strictly in this JSON format:
-            
             {{
               "classification": "bleak | neutral | positive",
               "reasoning": "short rationale here"
             }}
-            
+
             Context:
             {context}
-            
+
             Question:
             {query}
             """
-            
+
             response = self.client.chat.completions.create(
-                model="gpt-4o-mini",  
+                model="gpt-4o-mini",
                 messages=[{"role": "user", "content": prompt.strip()}],
                 temperature=0.3,
                 max_tokens=300
             )
-            
+
             raw_text = response.choices[0].message.content.strip()
-            
-            try:
-                result = json.loads(raw_text)
-                return result["classification"], result["reasoning"]
-            except json.JSONDecodeError:
-                print("⚠️ Failed to parse LLM response as JSON")
-                return majority_context, "Based on document similarity"
-                
+            result = json.loads(raw_text)
+            return result["classification"], result["reasoning"]
+
         except Exception as e:
-            print(f"❌ Error in LLM classification: {e}")
+            print(f"❌ LLM classification failed: {e}")
             return self.classify_context_rule_based(query)
-    
+
     def classify_context_rule_based(self, query):
-        """Rule-based fallback when LLM is unavailable"""
-        # Simple keyword-based classification
+        """Fallback classification based on simple keyword rules"""
         bleak_keywords = ["recession", "layoffs", "downturn", "crisis", "decline", "collapse"]
         positive_keywords = ["growth", "boom", "thriving", "expansion", "upward", "prosperity"]
-        
+
         query_lower = query.lower()
-        
-        bleak_count = sum(1 for word in bleak_keywords if word in query_lower)
-        positive_count = sum(1 for word in positive_keywords if word in query_lower)
-        
+        bleak_count = sum(word in query_lower for word in bleak_keywords)
+        positive_count = sum(word in query_lower for word in positive_keywords)
+
         if bleak_count > positive_count:
             return "bleak", "Based on negative economic indicators in query"
         elif positive_count > bleak_count:
             return "positive", "Based on positive economic indicators in query"
         else:
             return "neutral", "No strong indicators in either direction"
-    
+
     def get_context(self, query):
-        """Main method to get economic context classification"""
+        """
+        Main method for economic context classification, enhanced with Context Awareness Gate (CAG).
+
+        This function determines whether external document retrieval is necessary for the given query.
+        It does so using a hybrid approach:
+
+        1. Uses the Context Awareness Gate (CAG) to evaluate whether retrieval is needed, based on
+            query-document similarity and optional LLM signal.
+        2. If CAG decides retrieval is unnecessary, it falls back to a lightweight rule-based classifier.
+        3. If retrieval is needed and an LLM is available, it classifies context using retrieved documents + LLM.
+        4. If LLM is unavailable, it falls back to rule-based classification.
+
+        Args:
+            query (str): The user query about economic conditions.
+
+        Returns:
+            dict: {
+                'classification': One of "bleak", "neutral", or "positive",
+                'reasoning': Explanation of how the classification was determined,
+                'timestamp': ISO timestamp of classification decision
+            }
+        """
+        # Step 1: Check with Context Awareness Gate if retrieval is necessary
+        if hasattr(self, "cag") and not self.cag.should_retrieve(query):
+            context, reason = self.classify_context_rule_based(query)
+            return {
+                "classification": context,
+                "reasoning": f"Query handled without retrieval – {reason}",
+                "timestamp": datetime.now().isoformat()
+            }
+
+        # Step 2: Retrieval is required
+        # If LLM is available, classify using retrieved context and LLM
         if self.use_llm:
-            context, reasoning = self.classify_context_with_llm(query)
+            context, reason = self.classify_context_with_llm(query)
         else:
-            context, reasoning = self.classify_context_rule_based(query)
-            
+            # Fallback: rule-based if LLM is not available
+            context, reason = self.classify_context_rule_based(query)
+
+        # Return classification with reasoning and timestamp
         return {
             "classification": context,
-            "reasoning": reasoning,
+            "reasoning": reason,
             "timestamp": datetime.now().isoformat()
         }
